Add support for version constraint prefixes
This should address #32, this is a common type of version string for Composer and NPM packages so would be great to handle.

This doesn't extend to supporting complex version constraints like `>=1.4,<2.0` for instance. I'm not sure what the best approach would be in those situations to be honest. This change as-is would at least allow a significant improvement for merging simple version constraints.