clarified confusing caveat
About the dependabot waiting...

It was confusing to me, because the word _This_ in the original text (_This is a function&hellip;_) ws ambiguous as to whether it is referring to the functionality of the Dependabot or the Action. I.e., I thought that the text was saying that it is not an Action (despite being in the GitHub Actions Marketplace), but rather a modifier on the Dependabot functionality.

The updated language avoids this potential confusion.