Fix version detection
The current version detection fails since it's not strict enough. For example,
given the PR title "Bump @fontsource/lato from 4.5.5 to 4.5.8": Both "from"
and "to" versions are parsed as "4.5.5.", because the word "to" is included
in the dependency name as well.

Therefore, the corresponding regexes have been adapted to enforce a
whitespace prior to the "from" and "to" terms.